requirements.txt
selenium
pdfkit
pyyaml
webdriver-manager
